requirements.txt
flask==3.0.3
newspaper3k==0.2.8
lxml[html_clean]==5.2.2
lxml_html_clean==0.4.1
nltk==3.8.1
langdetect==1.0.9
transformers==4.41.2
torch==2.3.1
sentencepiece==0.2.0
sacremoses==0.1.1
gunicorn==22.0.0
requests==2.32.3
Pillow==10.3.0
